USB subsystem update
Adds among other things transmit time estimation and splitting transfers
if they get too close to the End Of Frame. The latter may of may not be
a good idea. For long Control transfers, that is a must have, bulk
transfers have it by default, for iso and irq it is forbidden. It seems
like there are some devices that do not support control transfers split
across EOF.