Fix linking and compiling on GCC 6
Starts to use the GOT which we did not handle, yet. Would not have been a
problem if it were placed in flash or initialized ram, but it was placed
in .bss area.

For the asm part, i could not find a constraint that would allow me to
directly generate a label for a jump.